+++ src/Makefile.am
@@ -271,44 +271,43 @@ crypto_libbitcoin_crypto_a_SOURCES = \
   crypto/hmac_sha512.h \
   crypto/ripemd160.cpp \
   crypto/ripemd160.h \
-  crypto/multihash/blake.c \
-  crypto/multihash/bmw.c \
-  crypto/multihash/cubehash.c \
-  crypto/multihash/echo.c \
-  crypto/multihash/fugue.c \
-  crypto/multihash/groestl.c \
-  crypto/multihash/hamsi.c \
-  crypto/multihash/haval.c \
-  crypto/multihash/jh.c \
-  crypto/multihash/keccak.c \
-  crypto/multihash/luffa.c \
-  crypto/multihash/sha2big.c \
-  crypto/multihash/shabal.c \
-  crypto/multihash/shavite.c \
-  crypto/multihash/simd.c \
-  crypto/multihash/skein.c \
-  crypto/multihash/sph_blake.h \
-  crypto/multihash/sph_bmw.h \
-  crypto/multihash/sph_cubehash.h \
-  crypto/multihash/sph_echo.h \
-  crypto/multihash/sph_fugue.h \
-  crypto/multihash/sph_groestl.h \
-  crypto/multihash/sph_hamsi.h \
-  crypto/multihash/sph_haval.h \
-  crypto/multihash/sph_jh.h \
-  crypto/multihash/sph_keccak.h \
-  crypto/multihash/sph_luffa.h \
-  crypto/multihash/sph_sha2.h \
-  crypto/multihash/sph_shabal.h \
-  crypto/multihash/sph_shavite.h \
-  crypto/multihash/sph_simd.h \
-  crypto/multihash/sph_skein.h \
-  crypto/multihash/sph_types.h \
-  crypto/multihash/sph_whirlpool.h \
-  crypto/multihash/whirlpool.c \
-  crypto/multihash/aes_helper.c \
-  crypto/algos/x11/x11hash.h \
-  crypto/algos/hmq1725/hashblock.h \
+  crypto/algos/hashlib/blake.c \
+  crypto/algos/hashlib/bmw.c \
+  crypto/algos/hashlib/cubehash.c \
+  crypto/algos/hashlib/echo.c \
+  crypto/algos/hashlib/fugue.c \
+  crypto/algos/hashlib/groestl.c \
+  crypto/algos/hashlib/hamsi.c \
+  crypto/algos/hashlib/haval.c \
+  crypto/algos/hashlib/jh.c \
+  crypto/algos/hashlib/keccak.c \
+  crypto/algos/hashlib/luffa.c \
+  crypto/algos/hashlib/sha2big.c \
+  crypto/algos/hashlib/shabal.c \
+  crypto/algos/hashlib/shavite.c \
+  crypto/algos/hashlib/simd.c \
+  crypto/algos/hashlib/skein.c \
+  crypto/algos/hashlib/sph_blake.h \
+  crypto/algos/hashlib/sph_bmw.h \
+  crypto/algos/hashlib/sph_cubehash.h \
+  crypto/algos/hashlib/sph_echo.h \
+  crypto/algos/hashlib/sph_fugue.h \
+  crypto/algos/hashlib/sph_groestl.h \
+  crypto/algos/hashlib/sph_hamsi.h \
+  crypto/algos/hashlib/sph_haval.h \
+  crypto/algos/hashlib/sph_jh.h \
+  crypto/algos/hashlib/sph_keccak.h \
+  crypto/algos/hashlib/sph_luffa.h \
+  crypto/algos/hashlib/sph_sha2.h \
+  crypto/algos/hashlib/sph_shabal.h \
+  crypto/algos/hashlib/sph_shavite.h \
+  crypto/algos/hashlib/sph_simd.h \
+  crypto/algos/hashlib/sph_skein.h \
+  crypto/algos/hashlib/sph_types.h \
+  crypto/algos/hashlib/sph_whirlpool.h \
+  crypto/algos/hashlib/whirlpool.c \
+  crypto/algos/hashlib/aes_helper.c \
+  crypto/algos/hashlib/multihash.h \
   crypto/algos/scrypt/scrypt.cpp \
   crypto/algos/scrypt/scrypt-sse2.cpp \
   crypto/algos/scrypt/scrypt.h \
