+++ src/crypto/algos/hashlib/echo.c
