+++ src/crypto/algos/hashlib/hamsi.c
