+++ src/crypto/algos/hashlib/hamsi_helper.c
