+++ src/crypto/algos/hashlib/haval_helper.c
