+++ src/crypto/algos/hashlib/jh.c
