+++ src/crypto/algos/hashlib/simd.c
