+++ src/crypto/algos/hashlib/sph_cubehash.h
