+++ src/crypto/algos/hashlib/sph_echo.h
