+++ src/crypto/algos/hashlib/sph_fugue.h
