+++ src/crypto/algos/hashlib/sph_hamsi.h
