+++ src/crypto/algos/hashlib/sph_haval.h
