+++ src/crypto/algos/hashlib/sph_jh.h
