+++ src/crypto/algos/hashlib/sph_types.h
