+++ src/primitives/block.h
@@ -51,6 +51,8 @@ public:
     int32_t nVersion;
     uint256 hashPrevBlock;
     uint256 hashMerkleRoot;
+	uint32_t nHeight;
+	uint32_t nReserved[7];
     uint32_t nTime;
     uint32_t nBits;
     uint32_t nNonce;
